# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "E-Learning Rest API DOCUMENTATION",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/quiz-attempts": {
            "get": {
                "tags": [
                    "Front Api Attempts"
                ],
                "summary": "get all attempts of student",
                "operationId": "c6c0e827f5c77e08dbabef1f70cf412b",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            }
        },
        "/api/quiz-attempts/{quizAttempt_id}": {
            "get": {
                "tags": [
                    "Front Api Attempts"
                ],
                "summary": "show quiz attempt and answers",
                "operationId": "59574a461f03d16af8c13661902d5fb5",
                "parameters": [
                    {
                        "name": "quizAttempt_id",
                        "in": "path",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/quiz-attempts/{quiz_id}": {
            "post": {
                "tags": [
                    "Front Api Quiz Attempt"
                ],
                "summary": "update user",
                "operationId": "a556d1e91e9da83e759196275d510074",
                "parameters": [
                    {
                        "name": "quiz_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "answers": {
                                        "type": "object",
                                        "example": {
                                            "10": 0,
                                            "25": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/quizzes/{quiz_id}": {
            "get": {
                "tags": [
                    "Front Api Quizzes"
                ],
                "summary": "show quiz",
                "operationId": "80ae73dd2a0d7af1c36dc59aac1ee1db",
                "parameters": [
                    {
                        "name": "quiz_id",
                        "in": "path",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        }
    }
}